UPDATE: retype float to double for all amount
将所有充值相关的 float 数据类型改为 double 以免精度问题导致金额误差：float 有效位数过低

+++ src/com/zz/sdk/demo/MainActivity.java
@@ -430,7 +430,7 @@ public class MainActivity extends Activity implements OnClickListener {
 			if (mDebugEnv != null) {
 				String str = ((TextView) findViewById(IDC_ET_RECHARGE_RATE))
 						.getText().toString().trim();
-				float rate;
+				double rate;
 				if (str.length() > 0) {
 					try {
 						rate = Float.parseFloat(str);
